+++ units/grs/staff_documents.py
@@ -1,6 +1,6 @@
 # === FILE HEADER ===
 # Title: Staff Documents
-# Path: grs/staff_documents.py
+# Path: units/grs/staff_documents.py
 # Created: 2026-07-14
 # Created by: docshamxo
 # Modified:
@@ -84,6 +84,6 @@ if __name__ == "__main__":
     send_grs_staff_documents()
 
 # === FILE FOOTER ===
-# End of file: grs/staff_documents.py
+# End of file: units/grs/staff_documents.py
 # Maintained by: docshamxo
 # === END FILE FOOTER ===