+++ units/ote/public_information.py
@@ -1,6 +1,6 @@
 # === FILE HEADER ===
 # Title: Public Information
-# Path: ote/public_information.py
+# Path: units/ote/public_information.py
 # Created: 2026-07-14
 # Created by: docshamxo
 # Modified:
@@ -82,6 +82,6 @@ if __name__ == "__main__":
     send_ote_public_information()
 
 # === FILE FOOTER ===
-# End of file: ote/public_information.py
+# End of file: units/ote/public_information.py
 # Maintained by: docshamxo
 # === END FILE FOOTER ===